+++ packages/snap-networks-utils/src/index.ts
@@ -3,6 +3,9 @@ export { createPrefixedLogger, logger, noOpLogger } from './logger';
 export { AssetsService } from './services/assets/AssetsService';
 export {
   ASSETS_SERVICE_NAME,
+  type AssetsControllerGetAccountAssetByIDAction,
+  type AssetsControllerGetAccountAssetsByIDsAction,
+  type AssetsControllerGetAccountAssetsByScopeAction,
   type AssetsServiceMessenger,
   type AssetsServiceMessengerCaller,
 } from './services/assets/messenger';
@@ -10,11 +13,9 @@ export type { AssetEntity, AssetScope } from './services/assets/types';
 export { mapControllerAsset } from './services/assets/utils/mapControllerAsset';
 export {
   CORE_MESSENGER_NAMESPACE,
-  type AssetsControllerGetAccountAssetByIDAction,
-  type AssetsControllerGetAccountAssetsByIDsAction,
-  type AssetsControllerGetAccountAssetsByScopeAction,
   type CoreMessenger,
   type CoreMessengerActions,
   type CoreMessengerCaller,
 } from './types/core-messenger';
+export type { MessengerCaller } from './types/messenger-caller';
 export { toUiAmount } from './utils/toUiAmount';